Delete force verify file and port command to other verify file
